requirements.txt
Flask~=3.0
Bootstrap-Flask~=2.3
Flask-Minify~=0.42
sqlalchemy~=2.0
flask-migrate~=4.0
alembic~=1.13
Flask-SQLAlchemy~=3.1
flask-login~=0.6
pytz~=2024.1
Flask-WTF~=1.2
WTForms~=3.1
email-validator~=2.1
PyJWT~=2.8
flask-mailman~=1.0
